[webapp] add telegramId resolver utility
